+++ R/utils.R
@@ -7,4 +7,38 @@
   }
 }
 
+# return character hostname from url
+get_hostname <- function(url) {
+  sub("^http[s]?://([^/:]+).*$", "\\1", url)
+}
+
+# return character or NULL, split string by comma
+parse_items <- function(x) {
+
+  vals <- strsplit(x, ",")[[1]]
+
+  vals
+}
+
+# return character regexes from no_proxy specs (parsed)
+regex_no_proxy <- function(no_proxy_parsed) {
+
+  # this function is vectorized
+
+  regex <- no_proxy_parsed
+
+  # if it starts with a dot, prepend with wildcard (*.)
+  regex <- sub("^[.]", "*.", regex)
+
+  # escape the existing dots: prepend all . with \\
+  regex <- gsub("\\.", "\\\\.", regex)
+
+  # where we see a wildcard (*), replace with (.*)
+  regex <- gsub("\\*", ".*", regex)
+
+  # prepend/append with ^/$
+  regex <- paste0("^", regex, "$")
+
+  regex
+}